Lin. Regression: Switched Student-T cdf code to computing series expansion from Abramowitz and Stegun. Gives way better results.

+++ methods/regress/src/pg_gp/student.h
@@ -10,6 +10,6 @@
 #ifndef STUDENT_H
 #define STUDENT_H
 
-extern double studentT_cdf(short /* nu */, double /* t */);
+extern double studentT_cdf(unsigned long /* nu */, double /* t */);
 
 #endif
